requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-multipart==0.0.20
pypdf==5.1.0
anthropic>=0.42.0
python-dotenv>=1.0.0
